Make FFI support optional (#2302)
